added some simple abundance tracking functionality
... in the form of arrays. unfortunately, as of now, every abundance,
etc. is read in individually; SDF can't handle arrays.

(I'm also not sure why the entire tree16 directory is considered as
modified by git, since it clearly wasn't. timestamp issue?)

+++ libmpmy/iozero.c
@@ -174,7 +174,7 @@ fseekrd0(int fd, long offset, int whence, void *buf, int reclen,
 		/* I'm not sure this is a well-defined operation */
 		doseek = (offset != 0);
 	    }else if( whence == MPMY_SEEK_SET ){
-		int cur_off = lseek(fd, 0, SEEK_CUR);
+		long cur_off = lseek(fd, 0, SEEK_CUR);
 		if (cur_off < 0){
 		    static int issued_seek_warning;
 		    if( errno == ESPIPE ){
